+++ src/main/java/fr/xephi/authme/task/purge/PurgeExecutor.java
@@ -7,7 +7,6 @@ import fr.xephi.authme.service.BukkitService;
 import fr.xephi.authme.service.PluginHookService;
 import fr.xephi.authme.settings.Settings;
 import fr.xephi.authme.settings.properties.PurgeSettings;
-import fr.xephi.authme.util.PlayerUtils;
 import org.bukkit.ChatColor;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.Server;
@@ -160,7 +159,7 @@ public class PurgeExecutor {
             makePath(settings.getProperty(PurgeSettings.DEFAULT_WORLD), "players"));
 
         for (OfflinePlayer offlinePlayer : cleared) {
-            File playerFile = new File(dataFolder, PlayerUtils.getUuidOrName(offlinePlayer) + ".dat");
+            File playerFile = new File(dataFolder, offlinePlayer.getUniqueId() + ".dat");
             if (playerFile.delete()) {
                 i++;
             }
@@ -192,7 +191,7 @@ public class PurgeExecutor {
 
         int deletedFiles = 0;
         for (OfflinePlayer offlinePlayer : cleared) {
-            File playerFile = new File(userDataFolder, PlayerUtils.getUuidOrName(offlinePlayer) + ".yml");
+            File playerFile = new File(userDataFolder, offlinePlayer.getUniqueId() + ".yml");
             if (playerFile.exists() && playerFile.delete()) {
                 deletedFiles++;
             }
